Empty signal handler for Windows
Save some resources.
CPU, memory: not measured
exe size: -4K
