+++ dnscrypt-proxy/hot_reload.go
@@ -1,15 +1,17 @@
 package main
 
 import (
-	"os"
-	"os/signal"
-	"syscall"
-
 	"github.com/jedisct1/dlog"
 )
 
 // InitHotReload sets up hot-reloading for configuration files
 func (proxy *Proxy) InitHotReload() error {
+	// Check if hot reload is enabled and platform has SIGHUP
+	if !proxy.enableHotReload && !HasSIGHUP {
+		dlog.Notice("Hot reload is disabled")
+		return nil
+	}
+
 	// Find plugins that support hot-reloading
 	plugins := []Plugin{}
 
@@ -104,28 +106,4 @@ func (proxy *Proxy) InitHotReload() error {
 	}
 
 	return nil
-}
-
-// setupSignalHandler sets up a SIGHUP handler to manually trigger reloads
-func setupSignalHandler(proxy *Proxy, plugins []Plugin) {
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, syscall.SIGHUP)
-
-	go func() {
-		for {
-			sig := <-sigChan
-			if sig == syscall.SIGHUP {
-				dlog.Notice("Received SIGHUP signal, reloading configurations")
-
-				// Reload each plugin that supports hot-reloading
-				for _, plugin := range plugins {
-					if err := plugin.Reload(); err != nil {
-						dlog.Errorf("Failed to reload plugin [%s]: %v", plugin.Name(), err)
-					} else {
-						dlog.Noticef("Successfully reloaded plugin [%s]", plugin.Name())
-					}
-				}
-			}
-		}
-	}()
 }
