+++ dnscrypt-proxy/signal_others.go
@@ -1,0 +1,10 @@
+//go:build windows || (js && wasm) || wasip1
+
+package main
+
+const HasSIGHUP = false
+
+// setupSignalHandler sets up a SIGHUP handler to manually trigger reloads
+func setupSignalHandler(proxy *Proxy, plugins []Plugin) {
+	return
+}
